requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.0
asn1crypto==1.5.1
asttokens==2.0.5
backcall==0.2.0
black==22.1.0
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.0.4
colorama==0.4.4
cryptography==37.0.4
cssselect2==0.6.0
cycler==0.11.0
decorator==5.1.1
dj-database-url==1.0.0
Django==4.0.2
django-ckeditor==6.4.2
django-crispy-forms==1.14.0
django-filter==21.1
django-heroku==0.3.1
django-js-asset==2.0.0
djangorestframework==3.13.1
executing==0.8.2
fonttools==4.33.3
future==0.18.2
google-api-core==2.8.2
google-api-python-client==2.53.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
gunicorn==20.1.0
html5lib==1.1
httplib2==0.20.4
idna==3.3
ipython==8.0.1
jedi==0.18.1
Jinja2==3.1.2
kiwisolver==1.4.3
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
numpy==1.23.0
oauth2client==4.1.3
oscrypto==1.3.0
packaging==21.3
pandas==1.4.3
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==9.1.1
platformdirs==2.5.1
prompt-toolkit==3.0.28
protobuf==4.21.2
psycopg2==2.9.3
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDrive==1.3.1
Pygments==2.11.2
pyHanko==0.13.2
pyhanko-certvalidator==0.19.5
pyparsing==3.0.9
PyPDF3==1.0.6
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.11
requests==2.28.1
rsa==4.8
scipy==1.8.1
seaborn==0.11.2
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
svglib==1.4.1
tinycss2==1.1.1
tomli==2.0.1
tqdm==4.64.0
traitlets==5.1.1
tzdata==2021.5
tzlocal==4.2
uritemplate==4.1.1
uritools==4.0.0
urllib3==1.26.10
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==6.2.0
xhtml2pdf==0.2.8
